move general tests to test_pipestat, remove some redundancy
