+++ tests/test_db_only_mode.py
@@ -5,7 +5,7 @@ from pipestat.const import *
 from .conftest import DB_URL
 
 from .conftest import SERVICE_UNAVAILABLE
-from sqlmodel import Session, SQLModel, create_engine
+from sqlmodel import SQLModel, create_engine
 
 from sqlmodel.main import default_registry
 
@@ -44,155 +44,6 @@ class TestDatabaseOnly:
                 )
             except Exception as e:
                 pytest.fail(f"Pipestat manager construction failed: {e})")
-
-    @pytest.mark.parametrize(
-        "val",
-        [
-            {"name_of_something": "test_name"},
-            {"number_of_things": 1},
-            {"percentage_of_things": 10.1},
-        ],
-    )
-    # TODO: need to test reporting of more complex types
-    def test_report(
-        self,
-        val,
-        config_file_path,
-        schema_file_path,
-    ):
-        with ContextManagerDBTesting(DB_URL):
-            psm = SamplePipestatManager(
-                schema_path=schema_file_path,
-                record_identifier="constant_record_id",
-                database_only=True,
-                config_file=config_file_path,
-            )
-            psm.report(record_identifier="constant_record_id", values=val, force_overwrite=True)
-            val_name = list(val.keys())[0]
-
-            assert psm.select_records(
-                filter_conditions=[
-                    {
-                        "key": val_name,
-                        "operator": "eq",
-                        "value": val[val_name],
-                    },
-                ],
-            )
-
-    @pytest.mark.parametrize(
-        "val",
-        [
-            {"name_of_something": "test_name"},
-            {"number_of_things": 1},
-            {"percentage_of_things": 10.1},
-        ],
-    )
-    @pytest.mark.parametrize("pipeline_type", ["project", "sample"])
-    def test_report_samples_and_project(
-        self,
-        val,
-        config_file_path,
-        schema_with_project_with_samples_without_status,
-        pipeline_type,
-    ):
-        with ContextManagerDBTesting(DB_URL):
-            psm = SamplePipestatManager(
-                schema_path=schema_with_project_with_samples_without_status,
-                record_identifier="constant_record_id",
-                database_only=True,
-                config_file=config_file_path,
-            )
-            val_name = list(val.keys())[0]
-            if pipeline_type is True:
-                if val_name in psm.cfg[SCHEMA_KEY].project_level_data:
-                    psm.report(
-                        values=val,
-                        force_overwrite=True,
-                        strict_type=False,
-                        # pipeline_type=pipeline_type,
-                    )
-                    assert psm.backend.select(filter_conditions=[(val_name, "eq", val[val_name])])
-                else:
-                    pass
-                    # assert that this would fail to report otherwise.
-            if pipeline_type == "sample":
-                if val_name in psm.cfg[SCHEMA_KEY].sample_level_data:
-                    psm.report(
-                        values=val,
-                        force_overwrite=True,
-                        strict_type=False,
-                        # pipeline_type=pipeline_type,
-                    )
-                    val_name = list(val.keys())[0]
-                    assert psm.backend.select(filter_conditions=[(val_name, "eq", val[val_name])])
-                else:
-                    pass
-                    # assert that this would fail to report otherwise.
-
-    @pytest.mark.parametrize(
-        "val",
-        [
-            {
-                "collection_of_images": [
-                    {
-                        "items": {
-                            "properties": {
-                                "prop1": {
-                                    "properties": {
-                                        "path": "pathstring",
-                                        "title": "titlestring",
-                                    }
-                                }
-                            }
-                        }
-                    }
-                ]
-            },
-            {"output_file": {"path": "path_string", "title": "title_string"}},
-            {
-                "output_image": {
-                    "path": "path_string",
-                    "thumbnail_path": "thumbnail_path_string",
-                    "title": "title_string",
-                }
-            },
-            {
-                "output_file_in_object": {
-                    "properties": {
-                        "prop1": {"properties": {"path": "pathstring", "title": "titlestring"}}
-                    }
-                }
-            },
-        ],
-    )
-    def test_complex_object_report(
-        self,
-        val,
-        config_file_path,
-        recursive_schema_file_path,
-    ):
-        with ContextManagerDBTesting(DB_URL):
-            REC_ID = "constant_record_id"
-            psm = SamplePipestatManager(
-                schema_path=recursive_schema_file_path,
-                record_identifier=REC_ID,
-                database_only=True,
-                config_file=config_file_path,
-            )
-            psm.report(
-                record_identifier=REC_ID, values=val, force_overwrite=True
-            )  # Force overwrite so that resetting the SQL DB is unnecessary.
-            val_name = list(val.keys())[0]
-            assert psm.select_records(
-                filter_conditions=[
-                    {
-                        "key": val_name,
-                        "operator": "eq",
-                        "value": val[val_name],
-                    }
-                ]
-            )
 
     @pytest.mark.parametrize(["rec_id", "res_id"], [("sample2", "number_of_things")])
     def test_select_invalid_filter_column__raises_expected_exception(
